requirements_conda.txt
tqdm
boto3
requests
regex
cudatoolkit=10.1
scikit-learn
pytorch
torchvision
torchtext
captum
tensorflow
